+++ src/views/BlogPage/components/BlogPostList/style.ts
@@ -19,3 +19,10 @@ export const BlogPostList = styled.div`
     gap: 36px;
   }
 `;
+
+export const SpinnerWrapper = styled.section`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin: 50px 0;
+`;
